+++ packages/components-theme-editor/src/Examples/Dashboard.tsx
@@ -75,7 +75,7 @@ export const FauxDashboard = () => (
 
       <SpaceVertical align="center" gap="xxsmall">
         <Heading as="h2">Some title here</Heading>
-        <Heading as="h3" variant="secondary">
+        <Heading as="h3" color="secondary">
           Some less important information
         </Heading>
       </SpaceVertical>