+++ packages/components/src/DataTable/BulkActions/BulkActions.tsx
@@ -62,7 +62,7 @@ const BulkActionsLayout: FC<BulkActionsProps> = ({
   }
 
   const selectedItemsText = (
-    <Span variant="secondary" fontSize="xsmall">
+    <Span color="secondary" fontSize="xsmall">
       {message}
     </Span>
   )
@@ -85,7 +85,7 @@ const BulkActionsLayout: FC<BulkActionsProps> = ({
     <div className={className}>
       <Menu>
         <MenuDisclosure>
-          <Button iconAfter="ArrowDown" size="xsmall">
+          <Button iconAfter="ArrowDropDown" size="xsmall">
             Bulk Actions
           </Button>
         </MenuDisclosure>