+++ packages/components/src/Popover/stories/ContentOverflow.tsx
@@ -46,7 +46,7 @@ export const ContentOverflow: FC = ({ children }) => (
         </PopoverContent>
       }
     >
-      <ButtonOutline iconAfter="ArrowDown" m="xxlarge">
+      <ButtonOutline iconAfter="ArrowDropDown" m="xxlarge">
         {children}
       </ButtonOutline>
     </Popover>