+++ packages/design-tokens/src/system/color/textColor.ts
@@ -24,54 +24,15 @@
 
  */
 
-import React from 'react'
-import { Story } from '@storybook/react/types-6-0'
-import { Span, SpanProps } from './Span'
+import { system } from 'styled-system'
 
-const Template: Story<SpanProps> = (args) => <Span {...args} />
-
-export const Basic = Template.bind({})
-Basic.args = {
-  children: 'Span Text',
+export interface TextColorProps {
+  color?: string
 }
 
-export const XXXXLarge = Template.bind({})
-XXXXLarge.args = {
-  ...Basic.args,
-  fontSize: 'xxxxlarge',
-}
-
-export const Bold = Template.bind({})
-Bold.args = {
-  ...Basic.args,
-  fontWeight: 'bold',
-}
-
-export const Color = Template.bind({})
-Color.args = {
-  ...Basic.args,
-  color: 'critical',
-}
-
-export const TextTransform = Template.bind({})
-TextTransform.args = {
-  ...Basic.args,
-  textTransform: 'uppercase',
-}
-
-export const Wrapped = Template.bind({})
-Wrapped.args = {
-  ...Basic.args,
-  breakword: true,
-}
-
-export const TextDecoration = Template.bind({})
-TextDecoration.args = {
-  ...Basic.args,
-  textDecoration: 'line-through',
-}
-
-export default {
-  component: Span,
-  title: 'Span',
-}
+export const textColor = system({
+  color: {
+    property: 'color',
+    scale: 'colors',
+  },
+})